+++ DatadogObjc/Sources/Tracing/Propagation/HTTPHeadersWriter+objc.swift
@@ -16,13 +16,23 @@ public class DDHTTPHeadersWriter: NSObject {
     }
 
     @objc
-    @available(*, deprecated, message: "This will be removed in future versions of the SDK. Use `init(sampleRate:)` instead.")
+    @available(*, deprecated, message: "This will be removed in future versions of the SDK. Use `init(samplingStrategy: .custom(sampleRate:))` instead.")
     public convenience init(samplingRate: Float) {
         self.init(sampleRate: samplingRate)
     }
 
     @objc
+    @available(*, deprecated, message: "This will be removed in future versions of the SDK. Use `init(samplingStrategy: .custom(sampleRate:))` instead.")
     public init(sampleRate: Float = 20) {
-        swiftHTTPHeadersWriter = HTTPHeadersWriter(sampleRate: sampleRate)
+        swiftHTTPHeadersWriter = HTTPHeadersWriter(
+            samplingStrategy: .custom(sampleRate: sampleRate)
+        )
+    }
+
+    @objc
+    public init(samplingStrategy: DDTraceSamplingStrategy) {
+        swiftHTTPHeadersWriter = HTTPHeadersWriter(
+            samplingStrategy: samplingStrategy.swiftType
+        )
     }
 }